backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Firebase
firebase-admin==6.2.0

# Environment and Configuration
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Data Processing (if needed by your planner)
pandas==2.1.3
numpy==1.24.3

# ML Libraries (if used by your planner)
scikit-learn==1.3.0
joblib==1.3.2

# OpenAI (if used in your planner)
openai==1.3.0

# HTTP Requests (if needed)
httpx==0.25.0
requests==2.31.0

# Validation
pydantic==2.5.0

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1